# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QUANT INDUSTRY V1 API
  description: |
    Institutional-grade quantitative trading platform API.
    
    ## Overview
    This API provides endpoints for:
    - Real-time market data and quotes
    - Trading signals and order management
    - ML/AI brain for signal generation
    - Risk management and monitoring
    - Portfolio tracking and analytics
    - Options flow and GEX analysis
    - Backtesting and strategy optimization
    
    ## Authentication
    Currently supports API key authentication via environment variables.
    Future: JWT-based auth for multi-tenant support.
    
    ## WebSocket
    Real-time data available via WebSocket at `/ws/unified`
    
  version: 1.0.0
  contact:
    name: QUANT INDUSTRY Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://localhost:8001
    description: Backend server (alternative)
...
paths:
...
  /api/signals/{signalId}/execute:
    post:
      tags: [trading]
      summary: Execute a signal
      operationId: executeSignal
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResult'
  /api/signals/{signalId}/dismiss:
    post:
      tags: [trading]
      summary: Dismiss a signal
      operationId: dismissSignal
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Signal dismissed
...
  /api/orders:
    get:
      tags: [trading]
      summary: Get all orders
      operationId: getOrders
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [trading]
      summary: Submit an order
      operationId: submitOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '200':
          description: Order submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResult'
  /api/orders/{orderId}:
    delete:
      tags: [trading]
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        filled_qty:
          type: number
        order_type:
          type: string
        status:
          type: string
        limit_price:
          type: number
        stop_price:
          type: number
        created_at:
          type: string
          format: date-time
        filled_at:
          type: string
          format: date-time
    OrderRequest:
      type: object
      required: [symbol, side, quantity, order_type]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        limit_price:
          type: number
        stop_price:
          type: number
        time_in_force:
          type: string
          enum: [day, gtc, ioc, fok]
          default: day
    OrderResult:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [submitted, filled, partial, rejected, cancelled]
        filled_qty:
          type: number
        filled_price:
          type: number
        message:
          type: string
    ExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        order_id:
          type: string
        message:
          type: string
...